+++ src/gvmd.c
@@ -555,8 +555,14 @@ accept_and_maybe_fork (int server_socket, sigset_t *sigmask_current)
     }
   sockaddr_as_str (&addr, client_address);
 
-  /* Fork a child to serve the client. */
-  pid = fork ();
+  /* Fork a child to serve the client.
+   *
+   * Use the default handlers for termination signals in the child.  This
+   * is required because the child calls 'system' and 'g_spawn_sync' in many
+   * places.  As the child waits for the spawned command, the child will
+   * not return to any code that checks termination_signal, so the child
+   * can't use the signal handlers inherited from the main process. */
+  pid = fork_with_handlers ();
   switch (pid)
     {
       case 0:
@@ -643,8 +649,10 @@ fork_connection_internal (gvm_connection_t *client_connection,
   struct sigaction action;
   gchar *auth_uuid;
 
-  /* Fork a child to use as scheduler client and server. */
+  /* Fork a child to use as scheduler/event client and server. */
 
+  /* This must 'fork' and not 'fork_with_handlers' so that the next fork can
+   * decide about handlers. */
   pid = fork ();
   switch (pid)
     {
@@ -687,7 +695,10 @@ fork_connection_internal (gvm_connection_t *client_connection,
 
   is_parent = 0;
 
-  pid = fork ();
+  /* As with accept_and_maybe_fork, use the default handlers for termination
+   * signals in the child.  This is required for signals to work when the
+   * child is waiting for spawns and forks. */
+  pid = fork_with_handlers ();
   switch (pid)
     {
       case 0:
@@ -791,6 +802,17 @@ fork_connection_internal (gvm_connection_t *client_connection,
         /* This process is returned as the child of
          * fork_connection_for_scheduler so that the returned parent can wait
          * on this process. */
+
+        if (scheduler)
+          {
+            /* When used for scheduling this parent process waits for the
+             * child.  That means it does not use the loops which handle
+             * termination_signal.  So we need to use the regular handlers
+             * for termination signals. */
+            setup_signal_handler (SIGTERM, SIG_DFL, 0);
+            setup_signal_handler (SIGINT, SIG_DFL, 0);
+            setup_signal_handler (SIGQUIT, SIG_DFL, 0);
+          }
 
         /** @todo Give the parent time to prepare. */
         gvm_sleep (5);
@@ -896,65 +918,6 @@ cleanup ()
 
   /* Delete pidfile if this process is the parent. */
   if (is_parent == 1) pidfile_remove ("gvmd");
-}
-
-/**
- * @brief Setup signal handler.
- *
- * Exit on failure.
- *
- * @param[in]  signal   Signal.
- * @param[in]  handler  Handler.
- * @param[in]  block    Whether to block all other signals during handler.
- */
-static void
-setup_signal_handler (int signal, void (*handler) (int), int block)
-{
-  struct sigaction action;
-
-  memset (&action, '\0', sizeof (action));
-  if (block)
-    sigfillset (&action.sa_mask);
-  else
-    sigemptyset (&action.sa_mask);
-  action.sa_handler = handler;
-  if (sigaction (signal, &action, NULL) == -1)
-    {
-      g_critical ("%s: failed to register %s handler",
-                  __func__, sys_siglist[signal]);
-      exit (EXIT_FAILURE);
-    }
-}
-
-/**
- * @brief Setup signal handler.
- *
- * Exit on failure.
- *
- * @param[in]  signal   Signal.
- * @param[in]  handler  Handler.
- * @param[in]  block    Whether to block all other signals during handler.
- */
-static void
-setup_signal_handler_info (int signal,
-                           void (*handler) (int, siginfo_t *, void *),
-                           int block)
-{
-  struct sigaction action;
-
-  memset (&action, '\0', sizeof (action));
-  if (block)
-    sigfillset (&action.sa_mask);
-  else
-    sigemptyset (&action.sa_mask);
-  action.sa_flags |= SA_SIGINFO;
-  action.sa_sigaction = handler;
-  if (sigaction (signal, &action, NULL) == -1)
-    {
-      g_critical ("%s: failed to register %s handler",
-                  __func__, sys_siglist[signal]);
-      exit (EXIT_FAILURE);
-    }
 }
 
 #ifndef NDEBUG
@@ -1101,7 +1064,11 @@ update_nvt_cache_retry ()
   setup_signal_handler (SIGCHLD, SIG_DFL, 0);
   while (1)
     {
-      pid_t child_pid = fork ();
+      pid_t child_pid;
+
+      /* No need to worry about fork_with_handlers, because
+       * fork_update_nvt_cache already did that. */
+      child_pid = fork ();
       if (child_pid > 0)
         {
           int status, i;
@@ -1142,7 +1109,6 @@ update_nvt_cache_retry ()
     }
 }
 
-
 /**
  * @brief Update the NVT cache in a child process.
  *
@@ -1176,7 +1142,7 @@ fork_update_nvt_cache ()
       return -1;
     }
 
-  pid = fork ();
+  pid = fork_with_handlers ();
   switch (pid)
     {
       case 0:
@@ -1186,7 +1152,10 @@ fork_update_nvt_cache ()
 
         /* Clean up the process. */
 
-        pthread_sigmask (SIG_SETMASK, &sigmask_current, NULL);
+        if (sigmask_normal)
+          pthread_sigmask (SIG_SETMASK, sigmask_normal, NULL);
+        else
+          pthread_sigmask (SIG_SETMASK, &sigmask_current, NULL);
         /** @todo This should happen via gmp, maybe with "cleanup_gmp ();". */
         cleanup_manage_process (FALSE);
         if (manager_socket > -1) close (manager_socket);