+++ src/manage.h
@@ -3718,9 +3718,6 @@ manage_optimize (GSList *, const gchar *, const gchar *);
 /* Signal management */
 
 int
-get_termination_signal ();
-
-int
 sql_cancel ();
 
 #endif /* not _GVMD_MANAGE_H */